Move inventory ui related implementations
