+++ spigot/src/main/kotlin/com/wolfyscript/viewportl/spigot/gui/inventoryui/BukkitInventoryGuiHolder.kt
@@ -16,12 +16,14 @@
  *     along with this program.  If not, see <https://www.gnu.org/licenses/>.
  */
 
-package com.wolfyscript.viewportl.spigot.gui.interaction
+package com.wolfyscript.viewportl.spigot.gui.inventoryui
 
 import com.wolfyscript.scafall.spigot.api.wrappers.wrap
 import com.wolfyscript.viewportl.common.gui.ViewRuntimeImpl
 import com.wolfyscript.viewportl.gui.GuiHolder
 import com.wolfyscript.viewportl.gui.Window
+import com.wolfyscript.viewportl.spigot.gui.inventoryui.interaction.ClickInteractionDetailsImpl
+import com.wolfyscript.viewportl.spigot.gui.inventoryui.interaction.DragInteractionDetailsImpl
 import org.bukkit.event.inventory.InventoryClickEvent
 import org.bukkit.event.inventory.InventoryCloseEvent
 import org.bukkit.event.inventory.InventoryDragEvent
